Renamed root.reducer to root-reducer.
